refactor(config-ui): adjust offline as a page
