+++ config-ui/src/error/components/index.ts
@@ -17,6 +17,5 @@
  */
 
 export * from './db-migrate';
-export * from './offline';
 export * from './default';
 export * from './bp-upgrade';
