+++ config-ui/src/pages/index.ts
@@ -20,3 +20,4 @@ export * from './project';
 export * from './connection';
 export * from './blueprint';
 export * from './pipeline';
+export * from './offline';
